driver:src: Fix Coverity issues
This patch fixes few coverity issues of type
PRINTF_ARGS ,OVERFLOW_BEFORE_WIDEN

+++ driver/src/stream_switch/xaie_ss.c
@@ -1319,7 +1319,7 @@ AieRC XAie_StrmSwDeterministicMergeConfig(XAie_DevInst *DevInst,
 	}
 
 	RegAddr |= (u64)(StrmMod->DetMerge->ConfigBase +
-		StrmMod->DetMerge->ArbConfigOffset * Arbitor) +
+		(u64)StrmMod->DetMerge->ArbConfigOffset * (u64)Arbitor) +
 		XAie_GetTileAddr(DevInst, Loc.Row, Loc.Col);
 	if(Position > 1U) {
 		RegAddr += 0x4U;
@@ -1402,7 +1402,7 @@ static AieRC _XAie_StrmSwDeterministicMergeCtrl(XAie_DevInst *DevInst,
 	}
 
 	RegAddr |= (u64)(StrmMod->DetMerge->EnableBase +
-		StrmMod->DetMerge->ArbConfigOffset * Arbitor) +
+		(u64)StrmMod->DetMerge->ArbConfigOffset * (u64) Arbitor) +
 		XAie_GetTileAddr(DevInst, Loc.Row, Loc.Col);
 	RegVal = XAie_SetField(Enable, StrmMod->DetMerge->Enable.Lsb,
 			StrmMod->DetMerge->Enable.Mask);
